% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{biclique.decompose}
\alias{biclique.decompose}
\title{Generate decompositions of multi-null exposure graphs given null hypothesis}
\usage{
biclique.decompose(
  Zrealized,
  hypothesis,
  controls = list(method = "greedy", mina = 10, num_randomizations = 2000)
)
}
\arguments{
\item{Zrealized}{A vector that gives the realization of treatment assignment. Its length should match \code{Yrealized}
and is the number of units in the experiment.}

\item{hypothesis}{A list that contains three functions specifyting the experiment design and null hypothesis.
See details for further illustration.}

\item{controls}{A list that contains settings for biclique decomposition and covariates adjustment.
By default it is \code{list(method="greedy", mina=10, num_randomizations=2000)}.
That is, by default it uses greedy decomposition algorithm with \code{mina=10}, and the number of
randomizations to perform is 2000. See details for further illustration.}
}
\value{
a list \code{MNE} of clique decomposition and specified \code{controls}.
Each element of \code{MNE} records one biclique decomposed from a multi-null exposure
graph and contains its focal units and focal assignments.
}
\description{
Generate decompositions of multi-null exposure graphs given null hypothesis
}
