% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats.R
\name{CI_grid}
\alias{CI_grid}
\title{Confidence interval by grid method}
\usage{
CI_grid(ci_lb, ci_ub, ci_dec)
}
\arguments{
\item{ci_lb}{Pre-determined lower bound of the confidence interval.}

\item{ci_ub}{Pre-determined upper bound of the confidence interval.}

\item{ci_dec}{Decimal of the confidence interval.}
}
\description{
Compute the CI by inverting the test using a grid method
}
