% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clique_diagnostic}
\alias{clique_diagnostic}
\title{Diagnostic for Randomized Experiment}
\usage{
clique_diagnostic(
  struc,
  p_group,
  p_indi_t,
  p_indi_nt,
  test_check,
  N,
  fmax = 25,
  NR = 500,
  Nx = 500
)
}
\arguments{
\item{struc}{Structure of the cluster. It should have two columns. The first column
represents group ID, and the second column is the individual ID. Please see the example.}

\item{p_group}{The probability of a group being selected as treated group.}

\item{p_indi_t}{The probability of an individual in the treated group being treated.}

\item{p_indi_nt}{The probability of an individual in the non-treated group being treated.}

\item{test_check}{If \code{TRUE}, it returns the number of focal assignments of the largest biclique
which for any of its focal assignments, does not contain only one type of exposure. This is necessary
for the clique test to work properly. If \code{FALSE}, then it just returns the number of focal
assignments of the largest biclique.}

\item{N}{The number of randomizations performed.}
}
\description{
For different number of units selected, it returns the approximate number of focal
assignments of the largest possible biclique of the null exposure graph that takes the
selected units as focal units if we draw \code{N} randomizations. It can serve
as a diagnostic for choosing \code{minr}, \code{minc} or \code{minass} in the
function \code{clique_test}.
}
