% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats.R
\name{gen_tstat}
\alias{gen_tstat}
\title{Generate test statistics}
\usage{
gen_tstat(G, type)
}
\arguments{
\item{G}{The \eqn{N\times N} adjacency matrix where \eqn{N} is the number of units. \code{G[i,j]} equals 1
iff there is an edge between units \code{i} and \code{j} in the network, and we set
all diagonal entries of \code{G} to be 0.}

\item{type}{The statistics to use. Currently support \code{"elc","score","htn"}
in Athey et al. (2018) Exact p-Values for Network Interference section 5.}
}
\description{
Generate default test statistics given a network structure.
}
\details{
In the following \eqn{G} is the adjacency matrix and \eqn{F} is the vector of focal indicator.
\itemize{
\item \code{"elc"} is the edge-level-contrast test statistic, which equals
\eqn{\frac{\sum_{i,j\neq i} F_i G_{ij} (1-F_j) Z_j Y_i^{obs}}{\sum_{i,j\neq i} F_i G_{ij} (1-F_j) Z_j} -
\frac{\sum_{i,j\neq i} F_i G_{ij} (1-F_j) (1-Z_j) Y_i^{obs}}{\sum_{i,j\neq i} F_i G_{ij} (1-F_j) (1-Z_j)
}}

\item \code{"score"} is a score statistic, which equals
\eqn{ \mathrm{cov}\left( Y_i^{obs}-\hat{\alpha}-\hat{\tau}_{d} Z_i,~ \sum_{j=1}^N Z_j \bar{G}_{ij} \bigg| \sum_{j=1}^N G_{ij} >0, F_i=1 \right)}.

\item \code{"htn"} is the has-treated-neighbor statistic, which equals
\eqn{\frac{1}{\mathrm{sd}_{Y_F^{obs}}\mathrm{sd}_{\mathrm{indicator}_F} } \frac{1}{N_F} \sum_{i:F_i=1}
(Y_i^{obs}-\bar{Y}_F^{obs})1\{\sum_j G_{ij} Z_j (1-F_j) >0 \} }
where \eqn{N_F} is the number of focal units, \eqn{\cdot_F} means restricting a variable to focal units only,
and \eqn{\bar{Y}_F^{obs}} is the sample average of \eqn{Y_F^{obs}}.
}
}
