% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{greedy_decom}
\alias{greedy_decom}
\title{Greedy Algorithm to Find a Biclique}
\usage{
greedy_decom(ne, num_ass)
}
\arguments{
\item{ne}{The null-exposure graph}

\item{num_ass}{Controls the minimum number of focal assignments in the biclique found
by the algorithm (algorithm runtime is sensitive to this value).}
}
\value{
A biclique of the input null-exposure graph \code{ne}.
}
\description{
This function returns a biclique in a null-exposure graph using a greedy algorithm. The
minimum number of focal assignments in the biclique is \code{num_ass}.
}
