% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{This is a function that simply prints hello, world!
it is very stupid, but whatever}
\usage{
hello(a)
}
\arguments{
\item{a}{A number}
}
\description{
This is a function that simply prints hello, world!
it is very stupid, but whatever
}
