% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{out_bassefeller}
\alias{out_bassefeller}
\title{Simulates the outcome vector based on the treatment assignment vector. The specific data generating procedure is adopted from Basse & Feller (2018).}
\usage{
out_bassefeller(
  N,
  K,
  Zobs,
  tau_main,
  sig_c = 0.1,
  sig_y = 0.5,
  taus = 0.7,
  taup = 1.5,
  mu00 = 2,
  equal = T
)
}
\arguments{
\item{N}{The number of observations (kids).}

\item{K}{The number of houses.}

\item{Zobs}{Observed treatment assignment vector.}

\item{tau_main}{The main effect.}

\item{sig_c}{Standard error on the causal effect.}

\item{sig_y}{Standard error on the outcome vector.}

\item{taus}{Spillover effect.}

\item{taup}{Primary effect.}

\item{mu00}{The effect on pure control units.}

\item{equal}{A binary parameter that determines the exact house structure. If equal=TRUE, every household has equal number of kids. If equal=FALSE, every household samples the number of kids at random.}
}
\value{
An outcome vector.
}
\description{
Simulates the outcome vector based on the treatment assignment vector. The specific data generating procedure is adopted from Basse & Feller (2018).
}
