% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{out_clique_decomposition_greedy}
\alias{out_clique_decomposition_greedy}
\title{Decomposing Null Exposure Graph Using the Greedy Algorithm}
\usage{
out_clique_decomposition_greedy(NEgraph, Zobs_id, num_ass, stop_at_Zobs)
}
\arguments{
\item{NEgraph}{The null-exposure graph}

\item{Zobs_id}{The index location of the observed assignment vector in \code{Z}, \code{Z_a}, and \code{Z_b}.}

\item{num_ass}{Controls the minimum number of focal assignments in the biclique found
by the algorithm (algorithm runtime is sensitive to this value).}

\item{stop_at_Zobs}{A Boolean indicating whether the decomposition algorithm should stop when the clique containing the observed assignment is found.}
}
\value{
If \code{stop_at_Zobs} is \code{TRUE}, a matrix representing the clique to condition upon.
If \code{stop_at_Zobs} is \code{FALSE}, a list containing the clique decomposition of the null-exposure graph.
}
\description{
Outputs a biclique decomposition of the null-exposure graph using the greedy algorithm
proposed above. The resulting decomposition paritions the assignment space,
while the unit space may overlap.
}
