% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{out_example_network}
\alias{out_example_network}
\title{Generates example two-dimensional network of 3 Gaussians.}
\usage{
out_example_network(num_units)
}
\arguments{
\item{num_units}{The number of units in the network.}
}
\value{
A list of \code{D}, a square matrix of all pairwise Euclidean distances between units, and \code{thepoints}, a matrix of dimension \code{num_units} by 2 of coordinates of the generated network units. (binary).
}
\description{
Generates example two-dimensional network of 3 Gaussians.
}
