% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{out_exp_ind}
\alias{out_exp_ind}
\title{Gives a range of indices that are exposed to treatments. See \code{out_Z_household}.}
\usage{
out_exp_ind(ii, lind, hind)
}
\arguments{
\item{ii}{the index of interest}

\item{lind}{the first index in each household}

\item{hind}{the last index in each household}
}
\value{
A vector of indices exposed to the treatment
}
\description{
Gives a range of indices that are exposed to treatments. See \code{out_Z_household}.
}
