% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{out_NEgraph}
\alias{out_NEgraph}
\title{Constructs the null exposure graph based on binary matrices describing exposure conditions in the null hypothesis}
\usage{
out_NEgraph(D_a, D_b)
}
\description{
Constructs the null exposure graph based on binary matrices describing exposure conditions in the null hypothesis
}
