% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{out_NEgraph_ex}
\alias{out_NEgraph_ex}
\title{Generate the multi-null exposure graph.}
\usage{
out_NEgraph_ex(Zsub, Z0, expos, tol = 1.5e-08)
}
\arguments{
\item{Zsub}{An integer from \code{1:length(Z0)} that represents a sample from \code{Z0}.}

\item{Z0}{A vector of column index that represents the remaining treatment.
Please see Sec. 8 of the paper for more details.}

\item{expos}{A three-way array with dimensions (number of units x number of randomizations x dimension of exposure).
Its entry records the exposure \code{f_i(z)} of units under treatments, where the exposure has more than one dimensions.}
}
\value{
A matrix of dimension (number of units x \code{length(Z0)}) which is the multi-null exposure graph with
respect to \code{Z0[Zsub]} and \code{Zsub}.
}
\description{
Generate the multi-null exposure graph.
}
