% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{out_NEgraph_multi}
\alias{out_NEgraph_multi}
\title{Generate the multi-null exposure graph.}
\usage{
out_NEgraph_multi(Zsub, Z0, Z, num_units, exposure_fn)
}
\arguments{
\item{Zsub}{An integer from \code{1:length(Z0)} that represents a sample from \code{Z0}.}

\item{Z0}{A vector of column index that represents the remaining treatment.
Please see Sec. 8 of the paper for more details.}

\item{Z}{A matrix of dimension (\code{num_units} x \code{num_randomizations+1}) that records the randomizations
of treatments for each unit.}

\item{num_units}{Number of units in the experiment, equals to the length of the outcome vector.}

\item{exposure_fn}{A function that specifies the equality of exposure
of a pair of individuals under different randomization.}
}
\value{
A matrix of dimension (\code{num_units} x \code{length(Z0)}) which is the multi-null exposure graph with
respect to \code{Z0[Zsub]} and \code{Z0}.
}
\description{
Generate the multi-null exposure graph.
}
