% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{out_treat_household}
\alias{out_treat_household}
\title{this function gives the exposure statuses for all observations}
\usage{
out_treat_household(housestruct, K1)
}
\arguments{
\item{housestruct}{The house structure vector that gives the number of kids in each household}

\item{K1}{The number of houses to treat.}
}
\value{
A list specifies treated houses and treated kids
}
\description{
this function gives the exposure statuses for all observations
}
