% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{out_Z}
\alias{out_Z}
\title{Returns sparse matrix \code{Z} of dimension (number of units x number of randomizations, i.e. assignments.)}
\usage{
out_Z(pi, num_randomizations)
}
\arguments{
\item{pi}{A \code{vector} of length number of units.}

\item{num_randomizations}{A \code{scalar} denoting the number of assignments desired.}
}
\value{
\code{Z} (sparse binary matrix).  The first column is the observed assignmenht.
}
\description{
Returns sparse matrix \code{Z} of dimension (number of units x number of randomizations, i.e. assignments.)
}
