% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{out_Zprime}
\alias{out_Zprime}
\title{Returns matrix \code{Z} of dimension (number of units x number of randomizations, i.e. assignments.)}
\usage{
out_Zprime(N, K, equal = T, numrand)
}
\arguments{
\item{N}{The number of observations (kids).}

\item{K}{The number of houses.}

\item{equal}{A binary parameter that determines the exact house structure. If equal=TRUE, every household has equal number of kids. If equal=FALSE, every household is constructed by sampling the number of kids at random.}

\item{numrand}{The number of randomizations}
}
\value{
A matrix of dimension (number of units x number of randomizations, i.e. assignments.) comprised of 0, denoting a unit is pure control under that particular assignment, 1 denoting a unit is a spillover under that particular assignment, and 2, denoting a unit is treated under that particular assignment.
}
\description{
Returns matrix \code{Z} of dimension (number of units x number of randomizations, i.e. assignments.)
}
