% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sparsify}
\alias{sparsify}
\title{Casts matrix \code{mat} into a sparse matrix.}
\usage{
sparsify(mat)
}
\arguments{
\item{mat}{A matrix to be sparsified}
}
\value{
A sparse matrix.
}
\description{
Casts matrix \code{mat} into a sparse matrix.
}
