% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats.R
\name{T_elc}
\alias{T_elc}
\title{The Edge-Level Contrast Statistic}
\usage{
T_elc(y, z, is_focal)
}
\description{
The Edge-Level Contrast Statistic in Athey et al. (2018). It equals
\eqn{\frac{\sum_{i,j\neq i} F_i G_{ij} (1-F_j) W_j Y_i^{obs}}{\sum_{i,j\neq i} F_i G_{ij} (1-F_j) W_j} -
\frac{\sum_{i,j\neq i} F_i G_{ij} (1-F_j) (1-W_j) Y_i^{obs}}{\sum_{i,j\neq i} F_i G_{ij} (1-F_j) (1-W_j)
}}. \eqn{G} is the adjacency matrix and \eqn{F} is the vector of focal indicator.
}
\seealso{
Athey et al. (2018) Exact p-Values for Network Interference, Equation 5.
}
