% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats.R
\name{T_score}
\alias{T_score}
\title{Score Test Statistic}
\usage{
T_score(y, z, is_focal)
}
\description{
The Score Test Statistic in Athey et al. (2018). It equals
\eqn{ \mathrm{cov}\left( Y_i^{obs}-\hat{\alpha}-\hat{\tau}_{d} W_i,~ \sum_{j=1}^N W_j \bar{G}_{ij} \bigg| \sum_{j=1}^N G_{ij} >0, F_i=1 \right)}.
\eqn{G} is the adjacency matrix and \eqn{F} is the vector of focal indicator.
}
\seealso{
Athey et al. (2018) Exact p-Values for Network Interference Equation 7.
}
