% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ate}
\alias{ate}
\title{Functions that returns difference in means between exposures \code{b} and \code{a}, coded as \code{1} and \code{-1}, respectively.}
\usage{
ate(Z, Y)
}
\description{
Functions that returns difference in means between exposures \code{b} and \code{a}, coded as \code{1} and \code{-1}, respectively.
}
