% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clique_test}
\alias{clique_test}
\title{The main randomization test function.}
\usage{
clique_test(Y, Z, Z_a, Z_b, Zobs_id, minr, minc, ...)
}
\arguments{
\item{Y}{The observed outcome vector.}

\item{Z}{A binary matrix of dimension (number of units x number of randomizations, i.e. assignments.) storing the assignment vectors.}

\item{Z_a}{A binary matrix with dimension (number of units x number of randomizations, i.e. assignments.)  Row i, column j of the matrix corresponds to whether a unit i is exposed to \code{a} under assignment j.  Please see example.}

\item{Z_b}{A binary matrix with (number of units x number of randomizations, i.e. assignments.)  Row i, column j of the matrix corresponds to whether a unit i is exposed to \code{b} under assignment j.  Please see example.}

\item{Zobs_id}{The index location of the observed assignment vector in \code{Z}, \code{Z_a}, and \code{Z_b}.}

\item{minr}{The minimum number of focal units included in the cliques (algorithm runtime is sensitive to this value).}

\item{minc}{The minimum number of focal assignment included in the cliques (algorithm runtime is sensitive to this value).}

\item{...}{Other stuff ...}
}
\value{
A list of items summarizing the randomization test.
}
\description{
The main randomization test function.
}
