% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintest.R
\name{clique_test}
\alias{clique_test}
\title{The generalized main randomization test function.}
\usage{
clique_test(
  Yrealized,
  Zrealized,
  design_fn,
  exposure_i,
  null_equiv,
  decomposition.control = list(method = "greedy", minass = 10),
  teststat,
  alpha = 0.05,
  Xadj = NULL,
  num_randomizations = 2000,
  ...
)
}
\arguments{
\item{Yrealized}{The observed outcome vector.}

\item{Zrealized}{A vector that gives the realization of treatment assignment. Its length should match \code{Yrealized}
and is the number of units in the experiment.}

\item{design_fn}{A function that returns a realization of treatment for the whole sample. For example,
if each unit has equal probability \eqn{0.2} to receive the treatment independently, we can write
\code{design_fn = function() { rbinom(num_units, 1, prob=0.2) }}.}

\item{exposure_i}{A function that returns exposure \eqn{f_i(z)} of unit \eqn{i} under treatment \eqn{z} where
\eqn{z} is the treatment for the whole sample. The inputs of the function are an index \code{i} and
a vector \code{z}. For example, if the exposure of \eqn{i} under \eqn{z} is the treatment it receives, then
we can write \code{exposure_i = function(z, i) { z[i] }}. See more examples in the README file.}

\item{null_equiv}{A function that takes two inputs from \code{exposure_i} and determines
whether \eqn{f_i(z_1)} is equivalent to \eqn{f_i(z_2)} under the null hypothesis. For example, if the
null is "extent of interference" type of null, we can write
\code{null_equiv = function(exposure_z1, exposure_z2) {identical(exposure_z1, exposure_z2)}}.}

\item{decomposition.control}{A list that contains settings for biclique decomposition. It should contains
a \code{method} which is either \code{"bimax"} or \code{"greedy"}. If \code{"bimax"} is used,
\code{minr} and \code{minc} should be supplied that specify the minimum number of units and assignments
in the bicliques found by the algorithm. If \code{"greedy"} is used, \code{minass} should be supplied.
By default, it is set to be \code{list(method="greedy", minass=10)}. That is, by default it uses
greedy decomposition algorithm with \code{minass=10}.}

\item{teststat}{The test statistic used. It should contain at least (with order)
\code{Y, Z, focal_units} as inputs where \code{Y} is the outcome vector, \code{Z} is the treatment vector
and \code{focal_units} is a 0-1 vector indicating whether a unit is focal (=1) or not (=0). All three inputs
should have length equal to number of units and have the same ordering.}

\item{alpha}{The significance level. By default it's \code{0.05}.}

\item{Xadj}{The covariates that might affect Y. If not \code{NULL}, will replace \code{Y} by the residuals
from the linear regression of \code{Y} on \code{Xadj}. Note that users would need to add an intercept to \code{Xadj} manually if they want.
To adjust \code{Xadj}, pass in \code{adj_Y=TRUE} and a non-empty \code{NULL} that has the same number of rows as \code{Y}.}

\item{num_randomizations}{Number of randomizations to perform. By default it's \code{2000}.}

\item{...}{Other stuff.}
}
\value{
A list of items summarizing the randomization test.
}
\description{
The generalized main randomization test function.
}
