% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintest.R
\name{clique_test}
\alias{clique_test}
\title{The generalized main randomization test function.}
\usage{
clique_test(
  Yrealized,
  Zrealized,
  design_fn,
  exposure_fn,
  exposure_i,
  null_equiv,
  decom = "greedy",
  alpha = 0.05,
  Xadj = NULL,
  num_randomizations = 2000,
  ...
)
}
\arguments{
\item{Yrealized}{The observed outcome vector.}

\item{Zrealized}{A vector of length being number of units that gives the realization of treatment assignment.}

\item{design_fn}{A function whose return is a one time realization of the distribution of the treatment.}

\item{exposure_fn}{A function that specifies the equality of exposure of a pair of individuals under different randomization.}

\item{decom}{The algorithm used to calculate the biclique decomposition. Currently supported algorithms are "bimax" and "greedy".}

\item{num_randomizations}{Number of randomizations to perform.}

\item{...}{Other stuff, such as parameters for clique decomposition algorithms (\code{minass} for
greedy decomposition, \code{minr} and \code{minc} for bimax decomposition).}
}
\value{
A list of items summarizing the randomization test.
}
\description{
The generalized main randomization test function.
}
