% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{out_house_structure}
\alias{out_house_structure}
\title{Generating household structure}
\usage{
out_house_structure(N = 900, K = N/3, equal = T)
}
\arguments{
\item{N}{The number of observations (kids)}

\item{K}{The number of households}

\item{equal}{A binary parameter that determines the exact house structure. If \code{equal=TRUE},
every household has equal number of kids.
If \code{equal=FALSE}, every household samples the number of kids at random. But in each household,
there would be at least two and at most \code{(2*N/K)-2} kids. Hence, on average there would be \code{N/K} kids per household.}
}
\value{
A vector of length \code{K} that gives the number of kids in each household.
}
\description{
Define a house structure vector that describes the population in each household,
which is useful in the clustered interference.
In this running example, houses refer to clusters and kids refer to experimental units.
}
