% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{out_NEgraph_multi_separate}
\alias{out_NEgraph_multi_separate}
\title{Generate multi-null exposure graph using null_equiv}
\usage{
out_NEgraph_multi_separate(Zsub, Z0, expos, null_equiv)
}
\arguments{
\item{Zsub}{An integer from \code{1:length(Z0)} that represents a sample from \code{Z0}.}

\item{Z0}{A vector of column index that represents the remaining treatment.
Please see Sec. 8 of the paper for more details.}

\item{expos}{A three-way array with dimensions (number of units x number of randomizations x dimension of exposure).
Its entry records the exposure \code{f_i(z)} of units under treatments, where the exposure has more than one dimensions.}

\item{null_equiv}{A function that determines whether (i,z1) is equivalent to (i,z2) under null.}
}
\value{
A matrix of dimension (\code{num_units} x \code{length(Z0)}) which is the multi-null exposure graph
with respect to \code{Z0[Zsub]} and \code{Z0}.
}
\description{
Generate multi-null exposure graph using null_equiv
}
