% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{out_treat_household}
\alias{out_treat_household}
\title{Generating exposure status for household}
\usage{
out_treat_household(housestruct, K1)
}
\arguments{
\item{housestruct}{The house structure vector that gives the number of kids in each household}

\item{K1}{The number of houses to treat.}
}
\value{
A list specifies treated houses and treated kids in each treated houses.
}
\description{
Outputs the exposure status for all observations in a household structure. Specifically, given
a household structure, it returns which households are to be treated, and within each treated
household, which kid is to be treated. Only one kid per treated household is going to be treated.
}
