% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{out_Yobs}
\alias{out_Yobs}
\title{Computes the observed outcome vector, given potential outcomes on the exposures}
\usage{
out_Yobs(Z, ya, yb)
}
\description{
Computes the observed outcome vector, given potential outcomes on the exposures
}
