% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_mod}
\alias{sample_mod}
\title{Get a sample from 1:k}
\usage{
sample_mod(k)
}
\arguments{
\item{k}{the maximum number that we can sample.}
}
\value{
An integer number sampled from 1 to \code{k}.
}
\description{
Get a sample from 1:k
}
